Move sections into components so that next lint will lint these files
