+++ frontend/components/sections/Focus.tsx
