+++ frontend/components/sections/Stats.tsx
