+++ frontend/components/sections/Values.tsx
